# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for getting geolocation info",
        "title": "Geolocation API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/geolocate/{ip}": {
            "get": {
                "description": "inquires IP geolocation's API to get details in json format",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "geolocate"
                ],
                "summary": "returns geolocation info given ip address",
                "parameters": [
                    {
                        "type": "string",
                        "description": "valid ipv4/v6 address",
                        "name": "ip",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid or missing ip address",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "unable to parse response from geolocation API",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}