# main.go
// GET /geolocate/<ip_address> IP geolocation that queries an IP address and returns location data, using a free service such as ipgeolocation.io
// GET /swagger/index.html an endpoint that displays Swagger/OpenAPI specification of the endpoint you created above. We recommend using echo-swagger for this.
// Finally, once you've generated your OpenAPI specification you will use Swagger codegen to generate a Golang SDK so that other Go applications can use and access your API. 
// Add the client code in a folder of your submission named /client.

package main

import (
  // "encoding/json"
  // "fmt"
  "os"
  "net"
  "html"
  "strings"
  "net/http"
  "io/ioutil"
  "github.com/k0kubun/pp"
  "github.com/labstack/echo/v4"
  "github.com/labstack/echo/v4/middleware"
  "github.com/swaggo/echo-swagger"
  _ "github.com/alvinlau/geoecho/docs"
)


// @title Geolocation API
// @version 1.0
// @description API for getting geolocation info
// @termsOfService http://swagger.io/terms/

// @contact.name API Support
// @contact.url http://www.swagger.io/support
// @contact.email support@swagger.io

// @license.name Apache 2.0
// @license.url http://www.apache.org/licenses/LICENSE-2.0.html

// @host localhost:8080
func main() {
  // Echo instance
  e := echo.New()

  // Middleware
  e.Use(middleware.Logger())
  e.Use(middleware.Recover())

  // Routes
  e.GET("/geolocate/:ip", geolocate)
  e.GET("/swagger/*", echoSwagger.WrapHandler)

  // Start server
  e.Logger.Fatal(e.Start(":" + os.Args[2]))
}

// Handlers

// geolocate godoc
// @Summary returns geolocation info given ip address
// @Description inquires IP geolocation's API to get details in json format
// @Tags geolocate
// @Accept plain
// @Produce json
// @Param ip path string true "valid ipv4/v6 address"
// @Success 200 {string} string "ok"
// @Failure 400 {string} string "invalid or missing ip address"
// @Failure 500 {string} string "unable to parse response from geolocation API"
// @Router /geolocate/{ip} [get]
func geolocate(c echo.Context) error {
  // parse the IP address
  ip := c.Param("ip") 
  pp.Println(ip)

  if strings.TrimSpace(ip) == "" || len(ip) == 0 {
    return c.String(http.StatusBadRequest, "missing ip address input")
  } else if net.ParseIP(html.UnescapeString(strings.ReplaceAll(ip, "%3A", ":"))) == nil {
    // https://golangbyexample.com/validate-an-ip-address-in-go/
    msg := "invalid ip address, make sure it is valid ip v4 or v6 address"
    return c.String(http.StatusBadRequest, msg)
  }

  // get API key
  apikey := os.Args[1]
  pp.Println(apikey)
  if strings.TrimSpace(apikey) == "" {
    return c.String(http.StatusInternalServerError, "API key not set")
  }

  // call ipgeolocation API
  resp, err := http.Get("https://api.ipgeolocation.io/ipgeo?apiKey=" + apikey + "&ip=" + ip)
  
  if err != nil {
    return c.String(http.StatusBadRequest, "invalid request to geolocation API")
  }
  defer resp.Body.Close()

  body, err := ioutil.ReadAll(resp.Body)
  if err != nil {
    return c.String(http.StatusInternalServerError, "unable to parse response from geolocation API")
  }
  return c.String(http.StatusOK, string(body))
}